+++ app/app.js
@@ -180,14 +180,13 @@ function App() {
 
     var userDatabaseService = new UserDatabaseService(metadataBackend);
 
-    var jobQueue = new JobQueue(metadataBackend);
     var jobPublisher = new JobPublisher(redis);
+    var jobQueue = new JobQueue(metadataBackend, jobPublisher);
     var userIndexer = new UserIndexer(metadataBackend);
-    var jobBackend = new JobBackend(metadataBackend, jobQueue, jobPublisher, userIndexer);
+    var jobBackend = new JobBackend(metadataBackend, jobQueue, userIndexer);
     var userDatabaseMetadataService = new UserDatabaseMetadataService(metadataBackend);
     var jobCanceller = new JobCanceller(userDatabaseMetadataService);
     var jobService = new JobService(jobBackend, jobCanceller);
-
 
     var genericController = new GenericController();
     genericController.route(app);
@@ -195,7 +194,7 @@ function App() {
     var queryController = new QueryController(userDatabaseService, tableCache, statsd_client);
     queryController.route(app);
 
-    var jobController = new JobController(userDatabaseService, jobService, jobCanceller);
+    var jobController = new JobController(userDatabaseService, jobService, statsd_client);
     jobController.route(app);
 
     var cacheStatusController = new CacheStatusController(tableCache);
@@ -210,7 +209,7 @@ function App() {
     var isBatchProcess = process.argv.indexOf('--no-batch') === -1;
 
     if (global.settings.environment !== 'test' && isBatchProcess) {
-        app.batch = batchFactory(metadataBackend);
+        app.batch = batchFactory(metadataBackend, statsd_client);
         app.batch.start();
     }