+++ app/controllers/job_controller.js
@@ -26,9 +26,10 @@ function getMaxSizeErrorMessage(sql) {
     );
 }
 
-function JobController(userDatabaseService, jobService) {
+function JobController(userDatabaseService, jobService, statsdClient) {
     this.userDatabaseService = userDatabaseService;
     this.jobService = jobService;
+    this.statsdClient = statsdClient || { increment: function () {} };
 }
 
 function bodyPayloadSizeMiddleware(req, res, next) {
@@ -101,17 +102,26 @@ JobController.prototype.cancelJob = function (req, res) {
                 return handleException(err, res);
             }
 
-            if ( req.profiler ) {
-                req.profiler.done('cancelJob');
-                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
-            }
-
             if (global.settings.api_hostname) {
-              res.header('X-Served-By-Host', global.settings.api_hostname);
+                res.header('X-Served-By-Host', global.settings.api_hostname);
             }
 
             if (result.host) {
-              res.header('X-Served-By-DB-Host', result.host);
+                res.header('X-Served-By-DB-Host', result.host);
+            }
+
+            if ( req.profiler ) {
+                req.profiler.done('cancelJob');
+                req.profiler.end();
+                req.profiler.sendStats();
+
+                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
+            }
+
+            if ( err ) {
+                self.statsdClient.increment('sqlapi.job.error');
+            } else {
+                self.statsdClient.increment('sqlapi.job.success');
             }
 
             res.send(result.job);
@@ -168,17 +178,26 @@ JobController.prototype.listJob = function (req, res) {
                 return handleException(err, res);
             }
 
-            if ( req.profiler ) {
-                req.profiler.done('listJob');
-                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
-            }
-
             if (global.settings.api_hostname) {
-              res.header('X-Served-By-Host', global.settings.api_hostname);
+                res.header('X-Served-By-Host', global.settings.api_hostname);
             }
 
             if (result.host) {
-              res.header('X-Served-By-DB-Host', result.host);
+                res.header('X-Served-By-DB-Host', result.host);
+            }
+
+            if ( req.profiler ) {
+                req.profiler.done('listJob');
+                req.profiler.end();
+                req.profiler.sendStats();
+
+                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
+            }
+
+            if ( err ) {
+                self.statsdClient.increment('sqlapi.job.error');
+            } else {
+                self.statsdClient.increment('sqlapi.job.success');
             }
 
             res.send(result.jobs);
@@ -234,17 +253,26 @@ JobController.prototype.getJob = function (req, res) {
                 return handleException(err, res);
             }
 
-            if ( req.profiler ) {
-                req.profiler.done('getJob');
-                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
-            }
-
             if (global.settings.api_hostname) {
-              res.header('X-Served-By-Host', global.settings.api_hostname);
+                res.header('X-Served-By-Host', global.settings.api_hostname);
             }
 
             if (result.host) {
-              res.header('X-Served-By-DB-Host', result.host);
+                res.header('X-Served-By-DB-Host', result.host);
+            }
+
+            if ( req.profiler ) {
+                req.profiler.done('getJob');
+                req.profiler.end();
+                req.profiler.sendStats();
+
+                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
+            }
+
+            if ( err ) {
+                self.statsdClient.increment('sqlapi.job.error');
+            } else {
+                self.statsdClient.increment('sqlapi.job.success');
             }
 
             res.send(result.job);
@@ -306,17 +334,26 @@ JobController.prototype.createJob = function (req, res) {
                 return handleException(err, res);
             }
 
-            if ( req.profiler ) {
-                req.profiler.done('persistJob');
-                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
-            }
-
             if (global.settings.api_hostname) {
                 res.header('X-Served-By-Host', global.settings.api_hostname);
             }
 
             if (result.host) {
                 res.header('X-Served-By-DB-Host', result.host);
+            }
+
+            if ( req.profiler ) {
+                req.profiler.done('persistJob');
+                req.profiler.end();
+                req.profiler.sendStats();
+
+                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
+            }
+
+            if ( err ) {
+                self.statsdClient.increment('sqlapi.job.error');
+            } else {
+                self.statsdClient.increment('sqlapi.job.success');
             }
 
             res.status(201).send(result.job);
@@ -378,17 +415,26 @@ JobController.prototype.updateJob = function (req, res) {
                 return handleException(err, res);
             }
 
-            if ( req.profiler ) {
-                req.profiler.done('updateJob');
-                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
-            }
-
             if (global.settings.api_hostname) {
-              res.header('X-Served-By-Host', global.settings.api_hostname);
+                res.header('X-Served-By-Host', global.settings.api_hostname);
             }
 
             if (result.host) {
-              res.header('X-Served-By-DB-Host', result.host);
+                res.header('X-Served-By-DB-Host', result.host);
+            }
+
+            if ( req.profiler ) {
+                req.profiler.done('updateJob');
+                req.profiler.end();
+                req.profiler.sendStats();
+
+                res.header('X-SQLAPI-Profiler', req.profiler.toJSONString());
+            }
+
+            if ( err ) {
+                self.statsdClient.increment('sqlapi.job.error');
+            } else {
+                self.statsdClient.increment('sqlapi.job.success');
             }
 
             res.send(result.job);