+++ batch/index.js
@@ -15,20 +15,19 @@ var JobBackend = require('./job_backend');
 var JobService = require('./job_service');
 var Batch = require('./batch');
 
-module.exports = function batchFactory (metadataBackend) {
+module.exports = function batchFactory (metadataBackend, statsdClient) {
     var queueSeeker = new QueueSeeker(metadataBackend);
     var jobSubscriber = new JobSubscriber(redis, queueSeeker);
-    var jobQueuePool = new JobQueuePool(metadataBackend);
     var jobPublisher = new JobPublisher(redis);
-    var jobQueue =  new JobQueue(metadataBackend);
+    var jobQueuePool = new JobQueuePool(metadataBackend, jobPublisher);
+    var jobQueue =  new JobQueue(metadataBackend, jobPublisher);
     var userIndexer = new UserIndexer(metadataBackend);
-    var jobBackend = new JobBackend(metadataBackend, jobQueue, jobPublisher, userIndexer);
+    var jobBackend = new JobBackend(metadataBackend, jobQueue, userIndexer);
     var userDatabaseMetadataService = new UserDatabaseMetadataService(metadataBackend);
-    // TODO: down userDatabaseMetadataService
-    var queryRunner = new QueryRunner();
+    var queryRunner = new QueryRunner(userDatabaseMetadataService);
     var jobCanceller = new JobCanceller(userDatabaseMetadataService);
     var jobService = new JobService(jobBackend, jobCanceller);
-    var jobRunner = new JobRunner(jobService, jobQueue, queryRunner, userDatabaseMetadataService);
+    var jobRunner = new JobRunner(jobService, jobQueue, queryRunner, statsdClient);
 
     return new Batch(jobSubscriber, jobQueuePool, jobRunner, jobService);
 };