+++ test/acceptance/job.fallback.test.js
@@ -15,6 +15,25 @@ var jobStatus = require('../../batch/job_status');
 
 describe('Batch API fallback job', function () {
 
+    function validateExpectedResponse(actual, expected) {
+        actual.query.forEach(function(actualQuery, index) {
+            var expectedQuery = expected.query[index];
+            assert.ok(expectedQuery);
+            Object.keys(expectedQuery).forEach(function(expectedKey) {
+                assert.equal(actualQuery[expectedKey], expectedQuery[expectedKey]);
+            });
+            var propsToCheckDate = ['started_at', 'ended_at'];
+            propsToCheckDate.forEach(function(propToCheckDate) {
+                if (actualQuery.hasOwnProperty(propToCheckDate)) {
+                    assert.ok(new Date(actualQuery[propToCheckDate]));
+                }
+            });
+        });
+
+        assert.equal(actual.onsuccess, expected.onsuccess);
+        assert.equal(actual.onerror, expected.onerror);
+    }
+
     var batch = batchFactory(metadataBackend);
 
     before(function () {
@@ -30,7 +49,6 @@ describe('Batch API fallback job', function () {
 
     describe('"onsuccess" on first query should be triggered', function () {
         var fallbackJob = {};
-
 
         it('should create a job', function (done) {
             assert.response(app, {
@@ -86,7 +104,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -133,7 +151,7 @@ describe('Batch API fallback job', function () {
                     "query": "SELECT * FROM untitle_table_4",
                     "onerror": "SELECT * FROM untitle_table_4 limit 1",
                     "status": "done",
-                    "fallback_status": "pending"
+                    "fallback_status": "skipped"
                 }]
             };
             var interval = setInterval(function () {
@@ -153,7 +171,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -221,7 +239,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.FAILED) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.DONE || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -268,7 +286,7 @@ describe('Batch API fallback job', function () {
                     query: 'SELECT * FROM nonexistent_table /* query should fail */',
                     onsuccess: 'SELECT * FROM untitle_table_4 limit 1',
                     status: 'failed',
-                    fallback_status: 'pending',
+                    fallback_status: 'skipped',
                     failed_reason: 'relation "nonexistent_table" does not exist'
                 }]
             };
@@ -290,7 +308,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.FAILED) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.DONE || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -358,7 +376,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.DONE && job.fallback_status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -424,9 +442,9 @@ describe('Batch API fallback job', function () {
                         return done(err);
                     }
                     var job = JSON.parse(res.body);
-                    if (job.status === jobStatus.FAILED && job.fallback_status === jobStatus.PENDING) {
+                    if (job.status === jobStatus.FAILED && job.fallback_status === jobStatus.SKIPPED) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -495,7 +513,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.FAILED && job.fallback_status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.DONE || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -560,9 +578,9 @@ describe('Batch API fallback job', function () {
                         return done(err);
                     }
                     var job = JSON.parse(res.body);
-                    if (job.status === jobStatus.DONE && job.fallback_status === jobStatus.PENDING) {
+                    if (job.status === jobStatus.DONE && job.fallback_status === jobStatus.SKIPPED) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -633,7 +651,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.DONE && job.fallback_status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -709,7 +727,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -759,13 +777,13 @@ describe('Batch API fallback job', function () {
                     "query": "SELECT * FROM nonexistent_table /* should fail */",
                     "onsuccess": "SELECT * FROM untitle_table_4 limit 1",
                     "status": "failed",
-                    "fallback_status": "pending",
+                    "fallback_status": "skipped",
                     "failed_reason": 'relation "nonexistent_table" does not exist'
                 }, {
                     "query": "SELECT * FROM untitle_table_4 limit 2",
                     "onsuccess": "SELECT * FROM untitle_table_4 limit 3",
-                    "status": "pending",
-                    "fallback_status": "pending"
+                    "status": "skipped",
+                    "fallback_status": "skipped"
                 }]
             };
 
@@ -786,7 +804,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.FAILED) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.DONE || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -842,7 +860,7 @@ describe('Batch API fallback job', function () {
                     "query": "SELECT * FROM nonexistent_table /* should fail */",
                     "onsuccess": "SELECT * FROM untitle_table_4 limit 3",
                     "status": "failed",
-                    "fallback_status": "pending",
+                    "fallback_status": "skipped",
                     "failed_reason": 'relation "nonexistent_table" does not exist'
                 }]
             };
@@ -864,7 +882,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.FAILED) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.DONE || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -875,6 +893,219 @@ describe('Batch API fallback job', function () {
         });
     });
 
+    describe('"onerror" should not be triggered for any query and "skipped"', function () {
+        var fallbackJob = {};
+
+        it('should create a job', function (done) {
+            assert.response(app, {
+                url: '/api/v2/sql/job?api_key=1234',
+                headers: {
+                    'Content-Type': 'application/x-www-form-urlencoded',
+                    'host': 'vizzuality.cartodb.com'
+                },
+                method: 'POST',
+                data: querystring.stringify({
+                    query: {
+                        query: [{
+                            query: "SELECT * FROM untitle_table_4 limit 1",
+                            onerror: "SELECT * FROM untitle_table_4 limit 2"
+                        }, {
+                            query: "SELECT * FROM untitle_table_4 limit 3",
+                            onerror: "SELECT * FROM untitle_table_4 limit 4"
+                        }]
+                    }
+                })
+            }, {
+                status: 201
+            }, function (res, err) {
+                if (err) {
+                    return done(err);
+                }
+                fallbackJob = JSON.parse(res.body);
+                done();
+            });
+        });
+
+        it('job should be failed', function (done) {
+            var expectedQuery = {
+                query: [{
+                    query: 'SELECT * FROM untitle_table_4 limit 1',
+                    onerror: 'SELECT * FROM untitle_table_4 limit 2',
+                    status: 'done',
+                    fallback_status: 'skipped'
+                }, {
+                    query: 'SELECT * FROM untitle_table_4 limit 3',
+                    onerror: 'SELECT * FROM untitle_table_4 limit 4',
+                    status: 'done',
+                    fallback_status: 'skipped'
+                }]
+            };
+
+            var interval = setInterval(function () {
+                assert.response(app, {
+                    url: '/api/v2/sql/job/' + fallbackJob.job_id + '?api_key=1234&',
+                    headers: {
+                        'Content-Type': 'application/x-www-form-urlencoded',
+                        'host': 'vizzuality.cartodb.com'
+                    },
+                    method: 'GET'
+                }, {
+                    status: 200
+                }, function (res, err) {
+                    if (err) {
+                        return done(err);
+                    }
+                    var job = JSON.parse(res.body);
+                    if (job.status === jobStatus.DONE) {
+                        clearInterval(interval);
+                        validateExpectedResponse(job.query, expectedQuery);
+                        done();
+                    } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
+                        clearInterval(interval);
+                        done(new Error('Job ' + job.job_id + ' is ' + job.status + ', expected to be done'));
+                    }
+                });
+            }, 50);
+        });
+    });
+
+    describe('"onsuccess" should be "skipped"', function () {
+        var fallbackJob = {};
+
+        it('should create a job', function (done) {
+            assert.response(app, {
+                url: '/api/v2/sql/job?api_key=1234',
+                headers: {
+                    'Content-Type': 'application/x-www-form-urlencoded',
+                    'host': 'vizzuality.cartodb.com'
+                },
+                method: 'POST',
+                data: querystring.stringify({
+                    query: {
+                        query: [{
+                            query: "SELECT * FROM untitle_table_4 limit 1, /* should fail */",
+                            onsuccess: "SELECT * FROM untitle_table_4 limit 2"
+                        }]
+                    }
+                })
+            }, {
+                status: 201
+            }, function (res, err) {
+                if (err) {
+                    return done(err);
+                }
+                fallbackJob = JSON.parse(res.body);
+                done();
+            });
+        });
+
+        it('job should be failed', function (done) {
+            var expectedQuery = {
+                query: [{
+                    query: 'SELECT * FROM untitle_table_4 limit 1, /* should fail */',
+                    onsuccess: 'SELECT * FROM untitle_table_4 limit 2',
+                    status: 'failed',
+                    fallback_status: 'skipped',
+                    failed_reason: 'syntax error at end of input'
+                }]
+            };
+
+            var interval = setInterval(function () {
+                assert.response(app, {
+                    url: '/api/v2/sql/job/' + fallbackJob.job_id + '?api_key=1234&',
+                    headers: {
+                        'Content-Type': 'application/x-www-form-urlencoded',
+                        'host': 'vizzuality.cartodb.com'
+                    },
+                    method: 'GET'
+                }, {
+                    status: 200
+                }, function (res, err) {
+                    if (err) {
+                        return done(err);
+                    }
+                    var job = JSON.parse(res.body);
+                    if (job.status === jobStatus.FAILED) {
+                        clearInterval(interval);
+                        validateExpectedResponse(job.query, expectedQuery);
+                        done();
+                    } else if (job.status === jobStatus.DONE || job.status === jobStatus.CANCELLED) {
+                        clearInterval(interval);
+                        done(new Error('Job ' + job.job_id + ' is ' + job.status + ', expected to be failed'));
+                    }
+                });
+            }, 50);
+        });
+    });
+
+
+    describe('"onsuccess" should not be triggered and "skipped"', function () {
+        var fallbackJob = {};
+
+        it('should create a job', function (done) {
+            assert.response(app, {
+                url: '/api/v2/sql/job?api_key=1234',
+                headers: {
+                    'Content-Type': 'application/x-www-form-urlencoded',
+                    'host': 'vizzuality.cartodb.com'
+                },
+                method: 'POST',
+                data: querystring.stringify({
+                    query: {
+                        query: [{
+                            query: "SELECT * FROM untitle_table_4 limit 1, /* should fail */",
+                        }],
+                        onsuccess: "SELECT * FROM untitle_table_4 limit 2"
+                    }
+                })
+            }, {
+                status: 201
+            }, function (res, err) {
+                if (err) {
+                    return done(err);
+                }
+                fallbackJob = JSON.parse(res.body);
+                done();
+            });
+        });
+
+        it('job should be failed', function (done) {
+            var expectedQuery = {
+                query: [{
+                    query: 'SELECT * FROM untitle_table_4 limit 1, /* should fail */',
+                    status: 'failed',
+                    failed_reason: 'syntax error at end of input'
+                }],
+                onsuccess: 'SELECT * FROM untitle_table_4 limit 2'
+            };
+
+            var interval = setInterval(function () {
+                assert.response(app, {
+                    url: '/api/v2/sql/job/' + fallbackJob.job_id + '?api_key=1234&',
+                    headers: {
+                        'Content-Type': 'application/x-www-form-urlencoded',
+                        'host': 'vizzuality.cartodb.com'
+                    },
+                    method: 'GET'
+                }, {
+                    status: 200
+                }, function (res, err) {
+                    if (err) {
+                        return done(err);
+                    }
+                    var job = JSON.parse(res.body);
+                    if (job.status === jobStatus.FAILED && job.fallback_status === jobStatus.SKIPPED) {
+                        clearInterval(interval);
+                        validateExpectedResponse(job.query, expectedQuery);
+                        done();
+                    } else if (job.status === jobStatus.DONE || job.status === jobStatus.CANCELLED) {
+                        clearInterval(interval);
+                        done(new Error('Job ' + job.job_id + ' is ' + job.status + ', expected to be failed'));
+                    }
+                });
+            }, 50);
+        });
+    });
 
     describe('"onsuccess" for first query should fail', function () {
         var fallbackJob = {};
@@ -942,7 +1173,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -1019,7 +1250,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -1097,7 +1328,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.DONE && job.fallback_status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -1177,7 +1408,7 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.DONE && job.fallback_status === jobStatus.DONE) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.FAILED || job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
@@ -1248,13 +1479,13 @@ describe('Batch API fallback job', function () {
                     var job = JSON.parse(res.body);
                     if (job.status === jobStatus.RUNNING && job.fallback_status === jobStatus.PENDING) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.status === jobStatus.DONE ||
                             job.status === jobStatus.FAILED ||
                             job.status === jobStatus.CANCELLED) {
                         clearInterval(interval);
-                        done(new Error('Job ' + job.job_id + ' is ' + job.status + ', expected to be done'));
+                        done(new Error('Job ' + job.job_id + ' is ' + job.status + ', expected to be running'));
                     }
                 });
             }, 50);
@@ -1266,7 +1497,7 @@ describe('Batch API fallback job', function () {
                     "query": "SELECT pg_sleep(3)",
                     "onsuccess": "SELECT pg_sleep(0)",
                     "status": "cancelled",
-                    "fallback_status": "pending"
+                    "fallback_status": "skipped"
                 }],
                 "onsuccess": "SELECT pg_sleep(0)"
             };
@@ -1285,8 +1516,8 @@ describe('Batch API fallback job', function () {
                     return done(err);
                 }
                 var job = JSON.parse(res.body);
-                if (job.status === jobStatus.CANCELLED && job.fallback_status === jobStatus.PENDING) {
-                    assert.deepEqual(job.query, expectedQuery);
+                if (job.status === jobStatus.CANCELLED && job.fallback_status === jobStatus.SKIPPED) {
+                    validateExpectedResponse(job.query, expectedQuery);
                     done();
                 } else if (job.status === jobStatus.DONE || job.status === jobStatus.FAILED) {
                     done(new Error('Job ' + job.job_id + ' is ' + job.status + ', expected to be cancelled'));
@@ -1355,7 +1586,7 @@ describe('Batch API fallback job', function () {
                     if (job.query.query[0].status === jobStatus.DONE &&
                         job.query.query[0].fallback_status === jobStatus.RUNNING) {
                         clearInterval(interval);
-                        assert.deepEqual(job.query, expectedQuery);
+                        validateExpectedResponse(job.query, expectedQuery);
                         done();
                     } else if (job.query.query[0].status === jobStatus.DONE ||
                             job.query.query[0].status === jobStatus.FAILED ||
@@ -1394,13 +1625,174 @@ describe('Batch API fallback job', function () {
                     return done(err);
                 }
                 var job = JSON.parse(res.body);
-                if (job.status === jobStatus.CANCELLED && job.fallback_status === jobStatus.PENDING) {
-                    assert.deepEqual(job.query, expectedQuery);
+                if (job.status === jobStatus.CANCELLED && job.fallback_status === jobStatus.SKIPPED) {
+                    validateExpectedResponse(job.query, expectedQuery);
                     done();
                 } else if (job.status === jobStatus.DONE || job.status === jobStatus.FAILED) {
                     done(new Error('Job ' + job.job_id + ' is ' + job.status + ', expected to be cancelled'));
                 }
             });
+        });
+    });
+
+    describe('should fail first "onerror" and job "onerror" and skip the other ones', function () {
+        var fallbackJob = {};
+
+        it('should create a job', function (done) {
+            assert.response(app, {
+                url: '/api/v2/sql/job?api_key=1234',
+                headers: {
+                    'Content-Type': 'application/x-www-form-urlencoded',
+                    'host': 'vizzuality.cartodb.com'
+                },
+                method: 'POST',
+                data: querystring.stringify({
+                    "query": {
+                        "query": [{
+                            "query": "SELECT * FROM atm_madrid limit 1, should fail",
+                            "onerror": "SELECT * FROM atm_madrid limit 2"
+                        }, {
+                            "query": "SELECT * FROM atm_madrid limit 3",
+                            "onerror": "SELECT * FROM atm_madrid limit 4"
+                        }],
+                        "onerror": "SELECT * FROM atm_madrid limit 5"
+                    }
+                })
+            }, {
+                status: 201
+            }, function (res, err) {
+                if (err) {
+                    return done(err);
+                }
+                fallbackJob = JSON.parse(res.body);
+                done();
+            });
+        });
+
+        it('job should fail', function (done) {
+            var expectedQuery = {
+                query: [{
+                    query: 'SELECT * FROM atm_madrid limit 1, should fail',
+                    onerror: 'SELECT * FROM atm_madrid limit 2',
+                    status: 'failed',
+                    fallback_status: 'failed',
+                    failed_reason: 'relation "atm_madrid" does not exist'
+                }, {
+                    query: 'SELECT * FROM atm_madrid limit 3',
+                    onerror: 'SELECT * FROM atm_madrid limit 4',
+                    status: 'skipped',
+                    fallback_status: 'skipped'
+                }],
+                onerror: 'SELECT * FROM atm_madrid limit 5'
+            };
+
+            var interval = setInterval(function () {
+                assert.response(app, {
+                    url: '/api/v2/sql/job/' + fallbackJob.job_id + '?api_key=1234&',
+                    headers: {
+                        'Content-Type': 'application/x-www-form-urlencoded',
+                        'host': 'vizzuality.cartodb.com'
+                    },
+                    method: 'GET'
+                }, {
+                    status: 200
+                }, function (res, err) {
+                    if (err) {
+                        return done(err);
+                    }
+                    var job = JSON.parse(res.body);
+                    if (job.status === jobStatus.FAILED && job.fallback_status === jobStatus.FAILED) {
+                        clearInterval(interval);
+                        validateExpectedResponse(job.query, expectedQuery);
+                        done();
+                    } else if (job.status === jobStatus.DONE || job.status === jobStatus.CANCELLED) {
+                        clearInterval(interval);
+                        done(new Error('Job ' + job.job_id + ' is ' + job.status + ', expected to be failed'));
+                    }
+                });
+            }, 50);
+        });
+    });
+
+    describe('should run first "onerror" and job "onerror" and skip the other ones', function () {
+        var fallbackJob = {};
+
+        it('should create a job', function (done) {
+            assert.response(app, {
+                url: '/api/v2/sql/job?api_key=1234',
+                headers: {
+                    'Content-Type': 'application/x-www-form-urlencoded',
+                    'host': 'vizzuality.cartodb.com'
+                },
+                method: 'POST',
+                data: querystring.stringify({
+                    "query": {
+                        "query": [{
+                            "query": "SELECT * FROM untitle_table_4 limit 1, should fail",
+                            "onerror": "SELECT * FROM untitle_table_4 limit 2"
+                        }, {
+                            "query": "SELECT * FROM untitle_table_4 limit 3",
+                            "onerror": "SELECT * FROM untitle_table_4 limit 4"
+                        }],
+                        "onerror": "SELECT * FROM untitle_table_4 limit 5"
+                    }
+                })
+            }, {
+                status: 201
+            }, function (res, err) {
+                if (err) {
+                    return done(err);
+                }
+                fallbackJob = JSON.parse(res.body);
+                done();
+            });
+        });
+
+        it('job should fail', function (done) {
+            var expectedQuery = {
+                "query": [
+                  {
+                    "query": "SELECT * FROM untitle_table_4 limit 1, should fail",
+                    "onerror": "SELECT * FROM untitle_table_4 limit 2",
+                    "status": "failed",
+                    "fallback_status": "done",
+                    "failed_reason": "LIMIT #,# syntax is not supported"
+                  },
+                  {
+                    "query": "SELECT * FROM untitle_table_4 limit 3",
+                    "onerror": "SELECT * FROM untitle_table_4 limit 4",
+                    "status": "skipped",
+                    "fallback_status": "skipped"
+                  }
+                ],
+                "onerror": "SELECT * FROM untitle_table_4 limit 5"
+            };
+
+            var interval = setInterval(function () {
+                assert.response(app, {
+                    url: '/api/v2/sql/job/' + fallbackJob.job_id + '?api_key=1234&',
+                    headers: {
+                        'Content-Type': 'application/x-www-form-urlencoded',
+                        'host': 'vizzuality.cartodb.com'
+                    },
+                    method: 'GET'
+                }, {
+                    status: 200
+                }, function (res, err) {
+                    if (err) {
+                        return done(err);
+                    }
+                    var job = JSON.parse(res.body);
+                    if (job.status === jobStatus.FAILED && job.fallback_status === jobStatus.DONE) {
+                        clearInterval(interval);
+                        validateExpectedResponse(job.query, expectedQuery);
+                        done();
+                    } else if (job.status === jobStatus.DONE || job.status === jobStatus.CANCELLED) {
+                        clearInterval(interval);
+                        done(new Error('Job ' + job.job_id + ' is ' + job.status + ', expected to be failed'));
+                    }
+                });
+            }, 50);
         });
     });
 });